requirements.txt
asgiref==3.6.0
attrs==22.2.0
backports.zoneinfo==0.2.1
black==23.1.0
click==8.1.3
Django==4.1.6
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1
Faker==18.7.0
flake8==6.0.0
gunicorn==20.1.0
importlib-resources==5.12.0
inflection==0.5.1
isort==5.12.0
jsonschema==4.17.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
pkgutil-resolve-name==1.3.10
platformdirs==3.2.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.5.0
uritemplate==4.1.1
zipp==3.15.0
